name refactors for internal packages, property, and suite
