+++ src/mlte/measurement/bind.py
@@ -4,7 +4,7 @@ Bind a measurement to one or more properties.
 
 from typing import Union
 
-from ..properties import Property
+from ..property import Property
 from .measurement import Measurement
