+++ src/mlte/measurement/memory/local_process_memory_consumption.py
@@ -9,7 +9,7 @@ from typing import Dict, Any
 from ..measurement import Measurement
 from ..evaluation import EvaluationResult
 from ..validation import Validator, Success, Failure
-from ...platform.os import is_windows
+from ..._private.platform import is_windows
 
 
 class MemoryStatistics(EvaluationResult):
