+++ src/mlte/property/costs/training_compute_cost.py
@@ -3,7 +3,7 @@ TrainingComputeCost property definition.
 """
 
 from ..property import Property
-from ...internal.text import cleantext
+from ..._private.text import cleantext
 
 
 class TrainingComputeCost(Property):
