+++ src/mlte/property/functionality/accuracy.py
@@ -3,7 +3,7 @@ Accuracy property definition.
 """
 
 from ..property import Property
-from ...internal.text import cleantext
+from ..._private.text import cleantext
 
 
 class Accuracy(Property):
