+++ test/measurement/cpu/test_local_process_cpu_utilization.py
@@ -9,7 +9,7 @@ import pytest
 import threading
 import subprocess
 
-from mlte.platform.os import is_windows, is_nix
+from mlte._private.platform import is_windows, is_nix
 from mlte.measurement.cpu import LocalProcessCPUUtilization
 
 from ...support.meta import path_to_support
